Remove gevent monkeypatching.
For somebody using this package as a library, it would be preferable to leave
whether to monkeypatch gevent into the standard library up to the user, rather
than forcing it on every consumer.

As far as I can tell, the only place that monkeypatching is used, is for
interacting with requests. The only place this is *directly* used by
the client is in `steam.core.CMClient.boostrap_from_webapi`, which is
only called on the initial connection. Letting that call be synchronous,
if the user doesn't want to use gevent directly seems reasonable.